# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Express label maker aopi",
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/hr/dpd/create/label": {
            "post": {
                "tags": [
                    "HrDpd"
                ],
                "summary": "Generira DPD naljepnicu za jednu narudžbu",
                "description": "Generira DPD naljepnicu za jednu narudžbu. Može vratiti 2 naljepnice ako je narudžba od 2 paketa. Parcel property prihvaća sve elemente iz DPD dokumentacije: [https://easyship.hr/documentation/hr/DPD_EasyShip_Web_Services.pdf#page=3](https://easyship.hr/documentation/hr/DPD_EasyShip_Web_Services.pdf#page=3)",
                "operationId": "ccc6a93a14ff26be5ec9df75fcc037e2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user": {
                                        "description": "User information",
                                        "properties": {
                                            "username": {
                                                "description": "Client’s Easyship username",
                                                "type": "string"
                                            },
                                            "password": {
                                                "description": "Client’s Easyship password",
                                                "type": "string"
                                            },
                                            "domain": {
                                                "description": "Client’s domain",
                                                "type": "string"
                                            },
                                            "licence": {
                                                "description": "Client’s ELM.com licence",
                                                "type": "string"
                                            },
                                            "email": {
                                                "description": "Client’s ELM.com E-mail address",
                                                "type": "string"
                                            },
                                            "platform": {
                                                "description": "WebStore platform",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "parcel": {
                                        "description": "Parcel information",
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Parcel label created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "parcels": {
                                                    "description": "Comma-separated parcel numbers",
                                                    "type": "string"
                                                },
                                                "label": {
                                                    "description": "Base64-encoded parcel label",
                                                    "type": "string",
                                                    "format": "base64"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid input or DPD Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "error_id": {
                                                        "description": "Error ID",
                                                        "type": "integer"
                                                    },
                                                    "error_details": {
                                                        "description": "Error details",
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}